requirements.txt
langchain==0.0.251
openai==0.27.7
python-dotenv==1.0.0
